WIP: T02 -- Split tests/unit/test_logging.py by subsystem
